♻️ Refactor Cvss Calculator for reusability
🔥 Remove unused class

🐛 Fix wrong element attrs
